requirements.txt
aiogram~=3.21.0
langchain~=0.3.26
python-dotenv~=1.1.1
reportlab~=4.4.2
sqlalchemy~=2.0.41